feat(user-ui)/split project components
