+++ packages/models/src/deliberation_sample_surveys/deliberation_sample_survey.rs
@@ -2,6 +2,7 @@
 use bdk::prelude::*;
 use validator::Validate;
 
+use crate::deliberation_response::DeliberationResponse;
 use crate::deliberation_user::DeliberationUser;
 use crate::SurveyV2;
 
@@ -44,4 +45,8 @@ pub struct DeliberationSampleSurvey {
     #[api_model(summary, many_to_many = deliberation_sample_survey_surveys, foreign_table_name = surveys, foreign_primary_key = survey_id, foreign_reference_key = sample_survey_id)]
     #[serde(default)]
     pub surveys: Vec<SurveyV2>,
+
+    #[api_model(summary, one_to_many = deliberation_responses, foreign_key = deliberation_id)]
+    #[serde(default)]
+    pub responses: Vec<DeliberationResponse>,
 }
