+++ packages/user-ui/src/pages/projects/_id/components/mod.rs
@@ -12,13 +12,8 @@ pub mod final_vote_modal;
 pub mod i18n;
 pub mod multiple_objective;
 pub mod my_final_survey;
-pub mod my_sample_survey;
 pub mod not_complete_survey_modal;
 pub mod project_header;
-pub mod remove_survey_modal;
-pub mod sample_statistics;
 pub mod sample_survey;
-pub mod sample_survey_info;
-pub mod sample_survey_question;
 pub mod single_objective;
 pub mod subjective;
