+++ packages/user-ui/src/pages/projects/_id/components/sample_survey/i18n.rs
@@ -1,0 +1,71 @@
+use bdk::prelude::*;
+
+translate! {
+    SampleSurveyTranslate;
+
+    title: {
+        ko: "표본 조사 주제",
+        en: "Sample Survey Title",
+    }
+    see_detail: {
+        ko: "자세히 보기",
+        en: "See Detail"
+    }
+    my_answer: {
+        ko: "내가 작성한 답변",
+        en: "My Answer"
+    }
+    response_per_question: {
+        ko: "질문별 응답",
+        en: "Responses to each question"
+    }
+    necessary: {
+        ko: "[필수]",
+        en: "[Necessary]"
+    }
+    plural: {
+        ko: "[복수]",
+        en: "[Plural]"
+    }
+    unit: {
+        ko: "명",
+        en: "Unit"
+    }
+    subjective_answer: {
+        ko: "주관식 답변",
+        en: "Subjective Answer"
+    }
+    update: {
+        ko: "수정하기",
+        en: "Update"
+    }
+    remove: {
+        ko: "삭제하기",
+        en: "Remove"
+    }
+    submit: {
+        ko: "제출하기",
+        en: "Submit"
+    }
+}
+
+translate! {
+    RemoveSurveyModalTranslate;
+
+    title: {
+        ko: "내가 작성한 답변을 삭제합니다.",
+        en: "I will delete the answer I wrote."
+    }
+    description: {
+        ko: "정말 삭제하시겠습니까?\n해당 표본조사를 삭제하면 모든 응답 데이터가 영구적으로 제거되며 복구할 수 없습니다.\n또한, 이 조사에 대한 보상 대상에서 제외됩니다.",
+        en: "Are you sure you want to delete it?\nIf you delete this survey, all response data will be permanently removed and cannot be recovered.\nYou will also no longer be eligible for compensation for this survey."
+    }
+    maintain: {
+        ko: "유지하기",
+        en: "Maintain"
+    }
+    remove: {
+        ko: "삭제",
+        en: "Remove"
+    }
+}
