+++ packages/user-ui/src/pages/projects/_id/components/sample_survey/remove_survey_modal.rs
@@ -1,5 +1,7 @@
 use bdk::prelude::*;
 
+use super::i18n::RemoveSurveyModalTranslate;
+
 #[component]
 pub fn RemoveSurveyModal(
     lang: Language,
@@ -29,26 +31,5 @@ pub fn RemoveSurveyModal(
                 }
             }
         }
-    }
-}
-
-translate! {
-    RemoveSurveyModalTranslate;
-
-    title: {
-        ko: "내가 작성한 답변을 삭제합니다.",
-        en: "I will delete the answer I wrote."
-    }
-    description: {
-        ko: "정말 삭제하시겠습니까?\n해당 표본조사를 삭제하면 모든 응답 데이터가 영구적으로 제거되며 복구할 수 없습니다.\n또한, 이 조사에 대한 보상 대상에서 제외됩니다.",
-        en: "Are you sure you want to delete it?\nIf you delete this survey, all response data will be permanently removed and cannot be recovered.\nYou will also no longer be eligible for compensation for this survey."
-    }
-    maintain: {
-        ko: "유지하기",
-        en: "Maintain"
-    }
-    remove: {
-        ko: "삭제",
-        en: "Remove"
     }
 }
