+++ packages/user-ui/src/pages/projects/_id/components/sample_survey/sample_survey_question.rs
@@ -6,11 +6,12 @@ use models::{response::Answer, Question, SurveyV2};
 use crate::{
     components::icons::left_arrow::LeftArrow,
     pages::projects::_id::components::{
-        multiple_objective::MultipleObjective, sample_survey::SampleSurveyTranslate,
-        single_objective::SingleObjective, subjective::Subjective,
+        multiple_objective::MultipleObjective, single_objective::SingleObjective,
+        subjective::Subjective,
     },
 };
 
+use super::i18n::SampleSurveyTranslate;
 #[component]
 pub fn SampleSurveyQuestion(
     lang: Language,
